Move definitions to .cpp
