+++ ydb/core/blobstorage/backpressure/load_based_timeout.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "defs.h"
 #include <algorithm>
 #include <atomic>
@@ -7,29 +9,11 @@ namespace NKikimr {
 // We assume only one active request can be in flight
 class TLoadBasedTimeoutManager {
 public:
-    TLoadBasedTimeoutManager(TDuration minimumTimeout, TDuration timePerRequestInFlight)
-        : MinimumTimeout(minimumTimeout)
-        , TimePerRequestInFlight(timePerRequestInFlight)
-        , ActiveRequest(false)
-    {}
+    TLoadBasedTimeoutManager(TDuration minimumTimeout, TDuration timePerRequestInFlight);
+    ~TLoadBasedTimeoutManager();
 
-    TDuration GetTimeoutForNewRequest() {
-        // if there is active request we don't increment counter
-        ui32 delta = std::exchange(ActiveRequest, true) ? 0 : 1;
-        ui32 requestsInFlight = RequestsInFlight.fetch_add(delta) + delta;
-        return std::max(MinimumTimeout, TimePerRequestInFlight * requestsInFlight);
-    }
-
-    void RequestCompleted() {
-        if (std::exchange(ActiveRequest, false)) {
-            ui32 prev = RequestsInFlight.fetch_sub(1);
-            Y_DEBUG_ABORT_UNLESS(prev > 0);
-        }
-    }
-
-    ~TLoadBasedTimeoutManager() {
-        RequestCompleted();
-    }
+    TDuration GetTimeoutForNewRequest();
+    void RequestCompleted();
 
 private:
     TDuration MinimumTimeout;
@@ -37,7 +21,5 @@ private:
     static std::atomic<ui32> RequestsInFlight;
     bool ActiveRequest;
 };
-
-std::atomic<ui32> TLoadBasedTimeoutManager::RequestsInFlight = 0;
 
 } // namespace NKikimr
