+++ ydb/core/blobstorage/backpressure/ya.make
@@ -16,6 +16,7 @@ SRCS(
     defs.h
     event.cpp
     event.h
+    load_based_timeout.cpp
     queue.cpp
     queue.h
     queue_backpressure_client.cpp
